+++ python/databases/inversereachability.py
@@ -66,13 +66,14 @@ from . import kinematicreachability, linkstatistics, inversekinematics
 import numpy
 import os.path
 from optparse import OptionParser
-try:
-    from scipy.optimize import leastsq
-except ImportError:
-    pass
 
 import logging
 log = logging.getLogger('openravepy.'+__name__.split('.',2)[-1])
+
+try:
+    from scipy.optimize import leastsq
+except ImportError:
+    print 'could not import scipy.optimize.leastsq'
 
 class InverseReachabilityModel(DatabaseGenerator):
     """Inverts the reachability and computes probability distributions of the robot's base given an end effector position"""
@@ -205,7 +206,7 @@ class InverseReachabilityModel(DatabaseGenerator):
             self.rotweight = heightthresh/quatthresh
             quateucdist2 = (1-cos(quatthresh))**2+sin(quatthresh)**2
             # find the density
-            basetrans = array(self.rmodel.reachabilitystats)
+            basetrans = array(self._GetValue(self.rmodel.reachabilitystats))
             assert len(basetrans) > 0
             basetrans[:,0:7] = poseMultArrayT(poseFromMatrix(Tbase),basetrans[:,0:7])
             # find the density of the points